Regex update to allow a tab as a delimiter for $ORIGIN and $INCLUDE directives
Some exported zone files (e.g., those provided by Omnis) are purely
tab-delimited.  The original regex match for $ORIGIN therefore failed
(expects a space character after the directive, but encountered a
tab).This appears to be allowable by the original RFC
[http://tools.ietf.org/html/rfc1035#section-5.1], so I changed the
regex matches for $ORIGIN and $INCLUDE to expect either a space or a
tab.